añadi funcion que verifica procesamiento de L2A

+++ bin/sentinel_L1CtoL2A.py
@@ -1,5 +1,5 @@
 import os
-from processing_sentinel2 import listaArchivos,nomDir,tipoCompresion,descomprime,obtieneFechaLog,obtieneAnio,sen2core,log,obtieneArchivoZip
+from processing_sentinel2 import listaArchivos,nomDir,tipoCompresion,descomprime,obtieneFechaLog,obtieneAnio,sen2core,log,obtieneArchivoZip,verificaLog
 
 #pathSen2core = '/home/lanotadm/sargazo/Sen2Cor-02.08.00-Linux64/bin/'
 #pathCFG = '/home/lanotadm/sen2cor/2.8/cfg/L2A_GIPP.xml'
@@ -20,28 +20,31 @@ for tile in tiles:
 			compresion = tipoCompresion(archivo)
 			if compresion == 'zip':
 				print(archivo)
-				descomprime(archivo,compresion,pathTmp)
-				dirL1C = nomDir(archivo,'L1C')
-				# Solucion para fallo de sen2core opcional
-				#os.system('sudo mkdir '+pathTmp+dirL1C+'/AUX_DATA')
-				sen2core(pathSen2core,pathCFG,pathTmp+dirL1C,pathOutput+tile+'/','20')
-				
-				#Comprimir el L2 y guardarlo en el txt
-				dirL2A = listaArchivos(pathTmp+'*L2*')[0]
-				anio = obtieneAnio(archivo)
-				archivoZip = obtieneArchivoZip(dirL2A)
+				if verificaLog(pathLog,archivo):
+					descomprime(archivo,compresion,pathTmp)
+					dirL1C = nomDir(archivo,'L1C')
+					# Solucion para fallo de sen2core opcional
+					#os.system('sudo mkdir '+pathTmp+dirL1C+'/AUX_DATA')
+					sen2core(pathSen2core,pathCFG,pathTmp+dirL1C,pathOutput+tile+'/','20')
+					
+					#Comprimir el L2 y guardarlo en el txt
+					dirL2A = listaArchivos(pathTmp+'*L2*')[0]
+					anio = obtieneAnio(archivo)
+					archivoZip = obtieneArchivoZip(dirL2A)
 
-				os.system('mkdir -p '+pathOutput+tile+'/'+anio)
-				
-				os.system('cd '+pathTmp+';zip -r '+pathOutput+tile+'/'+anio+'/'+dirL2A.split('/')[-1].split('.')[0]+'.zip ./'+dirL2A.split('/')[-1])
-				os.system('cd '+'/'.join(dirL2A.split('/')[:-1])+';zip -r '+archivoZip+' '+dirL2A.split('/')[-1]+';mv '+archivoZip+' '+pathOutput+tile+'/'+anio)
-				
-				os.system('rm -r '+pathTmp+dirL1C)
-				os.system('rm -r '+dirL2A)
-				#os.system('rm -r '+pathTmp+'\*L2*')
+					os.system('mkdir -p '+pathOutput+tile+'/'+anio)
+					
+					os.system('cd '+pathTmp+';zip -r '+pathOutput+tile+'/'+anio+'/'+dirL2A.split('/')[-1].split('.')[0]+'.zip ./'+dirL2A.split('/')[-1])
+					os.system('cd '+'/'.join(dirL2A.split('/')[:-1])+';zip -r '+archivoZip+' '+dirL2A.split('/')[-1]+';mv '+archivoZip+' '+pathOutput+tile+'/'+anio)
+					
+					os.system('rm -r '+pathTmp+dirL1C)
+					os.system('rm -r '+dirL2A)
+					#os.system('rm -r '+pathTmp+'\*L2*')
 
-				fecha = obtieneFechaLog()
-				
-				log(pathLog,archivo,pathOutput+tile+'/'+anio+'/'+dirL2A.split('/')[-1].split('.')[0]+'.zip',fecha)
-				#cont+=1
+					fecha = obtieneFechaLog()
+					
+					log(pathLog,archivo,pathOutput+tile+'/'+anio+'/'+dirL2A.split('/')[-1].split('.')[0]+'.zip',fecha)
+					#cont+=1
+				else:
+					print('Archivo: '+archivo+' ya fue procesado a L2A')
 #print(cont)
